moved fix to dedicated hook
